Add files via upload
There are four codes
Sum_Linked List
The program defines a ListNode struct to represent each node in the linked list, with each node having an integer value (val) and a pointer to the next node (next). It also defines a Solution class that provides the implementation for the addTwoNumbers function, which takes two linked lists l1 and l2 as input and returns the result of adding them together as a new linked list.

The addTwoNumbers function works by iterating through both linked lists simultaneously and adding the values of the nodes at each position together, taking care to handle any carry-over from previous positions. The result is stored in a new linked list represented by a ListNode structure.

The program also includes a helper function printList to print the contents of a linked list, which is used to display the result of adding the two numbers together.

In the main function, two linked lists l1 and l2 are created to represent the numbers to be added, and then the addTwoNumbers function is called to obtain the sum of these two numbers. The resulting linked list is printed using the printList function, and then all dynamically-allocated memory is freed before the program exits.
=================================
Two Sum problem
Already Defined . check TwoSum.cpp file
================================
Find Median Sorted Arrays
Also defined in FindMedianSortedArrays.cpp.
=================================
Palindrome
Already defined check the file